move get_description
